+++ packages/create-sitecore-jss/src/common/processes/transform.test.ts
@@ -1,6 +1,6 @@
 /* eslint-disable no-unused-expressions */
 import fs from 'fs-extra';
-import path from 'path';
+import path, { sep } from 'path';
 import chalk from 'chalk';
 import inquirer from 'inquirer';
 import ejs from 'ejs';
@@ -10,6 +10,7 @@ import sinon, { SinonStub } from 'sinon';
 import { currentPkg, partialPkg } from '../test-data/pkg';
 import * as transform from './transform';
 import * as helpers from '../utils/helpers';
+import { populateEjsData } from './transform';
 
 const {
   transformFilename,
@@ -464,6 +465,23 @@ describe('transform', () => {
     });
   });
 
+  describe('populateEjsData', () => {
+    it('should populate relative proxy path (with trailing slash) in helper, if proxyAppDestination populated', () => {
+      const answers = {
+        appName: 'JssNextWeb',
+        hostName: 'http://jssnextweb',
+        destination: 'samples/next',
+        proxyAppDestination: 'samples/proxy',
+        fetchWith: 'REST',
+        force: false,
+        templates: [],
+        language: 'en',
+      };
+      const result = populateEjsData(answers);
+      expect(result.helper.relativeProxyAppDestination).to.equal(`..${sep}proxy${sep}`);
+    });
+  });
+
   describe('transform', () => {
     let fsMkdirsSyncStub: SinonStub;
     let fsCopySyncStub: SinonStub;
@@ -476,7 +494,7 @@ describe('transform', () => {
     let diffAndWriteFilesStub: SinonStub;
     let writeFileToPathStub: SinonStub;
     let transformFilenameStub: SinonStub;
-    let openPackageJsonStub: SinonStub;
+    let openJsonFileStub: SinonStub;
     let log: SinonStub;
 
     beforeEach(() => {
@@ -495,7 +513,7 @@ describe('transform', () => {
       diffAndWriteFilesStub?.restore();
       writeFileToPathStub?.restore();
       transformFilenameStub?.restore();
-      openPackageJsonStub?.restore();
+      openJsonFileStub?.restore();
       log?.restore();
     });
 
@@ -560,6 +578,7 @@ describe('transform', () => {
           isDev: false,
           getPascalCaseName: helpers.getPascalCaseName,
           getAppPrefix: helpers.getAppPrefix,
+          relativeProxyAppDestination: `..${sep}proxy${sep}`,
         },
       });
       expect(diffAndWriteFilesStub).to.have.been.calledOnceWith({
@@ -663,7 +682,7 @@ describe('transform', () => {
 
       globSyncStub = sinon.stub(glob, 'sync').returns([file]);
       fsExistsSyncStub = sinon.stub(fs, 'existsSync').returns(true);
-      openPackageJsonStub = sinon.stub(helpers, 'openPackageJson').returns(currentPkg);
+      openJsonFileStub = sinon.stub(helpers, 'openJsonFile').returns(currentPkg);
       ejsRenderFileStub = sinon.stub(ejs, 'renderFile').returns(Promise.resolve(renderFileOutput));
       mergeStub = sinon.stub(transform, 'merge').returns(mergedPkg);
       diffAndWriteFilesStub = sinon.stub(transform, 'diffAndWriteFiles');
@@ -686,6 +705,48 @@ describe('transform', () => {
         },
       });
       expect(mergeStub).to.have.been.calledOnceWith(currentPkg, templatePkg);
+      expect(diffAndWriteFilesStub).to.have.been.calledOnceWith({
+        rendered: JSON.stringify(mergedPkg, null, 2),
+        pathToNewFile: path.join(destinationPath, file),
+        answers,
+      });
+    });
+
+    it('should merge json file', async () => {
+      const templatePath = path.resolve('templates/next');
+      const destinationPath = path.resolve('samples/next');
+      const file = 'test.json';
+      const renderFileOutput = '{ "one": 1, "two": 2}';
+      const currentJson = { three: 3, four: 4 };
+      const templateJson = JSON.parse(renderFileOutput);
+      const mergedPkg = { merged: true };
+
+      globSyncStub = sinon.stub(glob, 'sync').returns([file]);
+      fsExistsSyncStub = sinon.stub(fs, 'existsSync').returns(true);
+      openJsonFileStub = sinon.stub(helpers, 'openJsonFile').returns(currentJson);
+      ejsRenderFileStub = sinon.stub(ejs, 'renderFile').returns(Promise.resolve(renderFileOutput));
+      mergeStub = sinon.stub(transform, 'merge').returns(mergedPkg);
+      diffAndWriteFilesStub = sinon.stub(transform, 'diffAndWriteFiles');
+
+      const answers = {
+        destination: destinationPath,
+        templates: [],
+        appPrefix: false,
+        force: false,
+      };
+
+      await transformFunc(templatePath, answers);
+
+      expect(ejsRenderFileStub).to.have.been.calledOnceWith(path.join(templatePath, file), {
+        ...answers,
+        helper: {
+          isDev: false,
+          getPascalCaseName: helpers.getPascalCaseName,
+          getAppPrefix: helpers.getAppPrefix,
+        },
+      });
+      expect(mergeStub).to.have.been.calledOnceWith(currentJson, templateJson);
+      expect(openJsonFileStub).to.have.been.calledOnceWith(`${destinationPath}${sep}${file}`);
       expect(diffAndWriteFilesStub).to.have.been.calledOnceWith({
         rendered: JSON.stringify(mergedPkg, null, 2),
         pathToNewFile: path.join(destinationPath, file),