+++ packages/create-sitecore-jss/src/common/processes/transform.ts
@@ -8,7 +8,7 @@ import inquirer from 'inquirer';
 import {
   getPascalCaseName,
   getAppPrefix,
-  openPackageJson,
+  openJsonFile,
   sortKeys,
   writeFileToPath,
   isDevEnvironment,
@@ -192,6 +192,27 @@ export const diffAndWriteFiles = async ({
   }
 };
 
+export const populateEjsData = (answers: BaseArgs, destination?: string) => {
+  // pass in helper to answers object
+  const ejsData: Data = {
+    ...answers,
+    helper: {
+      isDev: isDevEnvironment(destination || answers.destination),
+      getPascalCaseName: getPascalCaseName,
+      getAppPrefix: getAppPrefix,
+    },
+  };
+  // When SPA application and XM Cloud proxy is used we need to calculate relative path between SPA app and proxy to apply that in EJS template
+  // Absolute path can't be used across all the machines
+  if (answers.proxyAppDestination) {
+    ejsData.helper.relativeProxyAppDestination = `${path.relative(
+      path.resolve(answers.destination),
+      path.resolve(answers.proxyAppDestination)
+    )}${sep}`;
+  }
+  return ejsData;
+};
+
 type TransformOptions = {
   /**
    * Determines whether a file should be copied only (not rendered through ejs)
@@ -232,7 +253,7 @@ export const transform = async (
   const { isFileForCopy, isFileForSkip, fileForCopyRegExp = FILE_FOR_COPY_REGEXP } = options;
   let destination = undefined;
   // allow proxy app to be installed separately alongside base app
-  if (templatePath.match(/.*node-.+-proxy$/g)) {
+  if (templatePath.match(/.*node-.+-proxy$/g) && answers.proxyAppDestination) {
     destination = answers.proxyAppDestination;
   }
   const destinationPath = path.resolve(destination || answers.destination);
@@ -241,16 +262,7 @@ export const transform = async (
     answers.appPrefix = false;
   }
 
-  // pass in helper to answers object
-  const ejsData: Data = {
-    ...answers,
-    helper: {
-      isDev: isDevEnvironment(destination || answers.destination),
-      getPascalCaseName: getPascalCaseName,
-      getAppPrefix: getAppPrefix,
-    },
-  };
-
+  const ejsData: Data = populateEjsData(answers, destination);
   // the templates to be run through ejs render or copied directly
   const files = glob.sync('**/*', { cwd: templatePath, dot: true, nodir: true });
 
@@ -282,13 +294,13 @@ export const transform = async (
         continue;
       }
 
-      if (file.endsWith('package.json') && fs.existsSync(pathToNewFile)) {
-        // we treat package.json a bit differently
-        // read the current package.json and the template package.json (rendered with ejs)
-        const currentPkg = openPackageJson(pathToNewFile);
-        const templatePkg = JSON.parse(await renderFile(path.resolve(pathToTemplate), ejsData));
+      if (file.endsWith('.json') && fs.existsSync(pathToNewFile)) {
+        // we treat a .json a bit differently
+        // read the current .json and the template .json (rendered with ejs)
+        const currentJson = openJsonFile(pathToNewFile);
+        const templateJson = JSON.parse(await renderFile(path.resolve(pathToTemplate), ejsData));
         // merge them and set the result to str which will then go through diff
-        const merged = merge(currentPkg, templatePkg);
+        const merged = merge(currentJson, templateJson);
         str = JSON.stringify(merged, null, 2);
       }