+++ packages/create-sitecore-jss/src/initializers/angular-xmcloud/index.ts
@@ -1,12 +1,7 @@
 import path, { sep } from 'path';
-import {
-  ClientAppArgs,
-  DEFAULT_APPNAME,
-  Initializer,
-  transform,
-  openPackageJson,
-} from '../../common';
+import { ClientAppArgs, DEFAULT_APPNAME, Initializer, transform, openJsonFile } from '../../common';
 import { InitializerResults } from '../../common/Initializer';
+import { getDefaultProxyDestination } from '../../common/utils/helpers';
 
 export default class AngularXmCloudInitializer implements Initializer {
   get isBase(): boolean {
@@ -14,8 +9,16 @@ export default class AngularXmCloudInitializer implements Initializer {
   }
 
   async init(args: ClientAppArgs) {
-    const pkg = openPackageJson(`${args.destination}${sep}package.json`);
+    const pkg = openJsonFile(`${args.destination}${sep}package.json`);
+    const addInitializers = [];
+    // when installing proxy alongside main app, have a separate path ready
+    if (!args.proxyAppDestination) {
+      args.proxyAppDestination = getDefaultProxyDestination(args.destination, 'node-xmcloud-proxy');
+    }
 
+    if (!args.templates.includes('node-xmcloud-proxy')) {
+      addInitializers.push('node-xmcloud-proxy');
+    }
     const mergedArgs = {
       ...args,
       appName: args.appName || pkg?.config?.appName || DEFAULT_APPNAME,
@@ -29,6 +32,7 @@ export default class AngularXmCloudInitializer implements Initializer {
     const response: InitializerResults = {
       nextSteps: [],
       appName: args.appName || DEFAULT_APPNAME,
+      initializers: addInitializers,
     };
 
     return response;