+++ packages/create-sitecore-jss/src/initializers/nextjs/remove-dev-dependencies.ts
@@ -1,14 +1,14 @@
 import fs from 'fs';
 import path from 'path';
-import { openPackageJson, writePackageJson } from '../../common';
+import { openJsonFile, writeJsonFile } from '../../common';
 
 export const removeDevDependencies = (projectPath: string) => {
   // remove `next-transpile-modules` dependency
   const packagePath = path.join(projectPath, 'package.json');
-  const pkg = openPackageJson(packagePath);
+  const pkg = openJsonFile(packagePath);
   if (pkg?.devDependencies['next-transpile-modules']) {
     delete pkg.devDependencies['next-transpile-modules'];
-    writePackageJson(pkg, packagePath);
+    writeJsonFile(pkg, packagePath);
   }
 
   // remove monorepo next.config.js plugin